+++ src/main/java/dev/lucasvital/auth/user/UserRepository.java
@@ -1,0 +1,8 @@
+package dev.lucasvital.auth.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+}